% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sliding_window}
\alias{sliding_window}
\title{Sliding Window function.}
\usage{
sliding_window(data, col, window_size, threshold_size)
}
\arguments{
\item{x}{A data set}

\item{y}{A column/vector in your data to manipulate formated as data$column}

\item{z}{Numeric: Sliding window size}

\item{q}{Numeric: Allowed threshold size}
}
\description{
Sliding Window function.
}
\examples{

vec1 <- rnorm(100)
vec2 <- runif(100)
vec3 <- rpois(100, lambda = 5)

df <- data.frame(column1 = vec1, column2 = vec2, column3 = vec3)

processed_data <- sliding_window(df, df$column1, window_size, threshold_size)
}
