% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sorting_speed}
\alias{sorting_speed}
\title{Sorting Speed Function}
\usage{
sorting_speed(array, y)
}
\arguments{
\item{array}{The input array to be sorted.}

\item{y}{Character vector specifying which sorting algorithms to evaluate.}
}
\value{
A list containing timing information and sorted arrays for each selected algorithm.
}
\description{
This function compares the speed of specified sorting algorithms on the given array. It allows you to insert a
numeric array, specify which sorting algorithms to compare and get output in a list format.
}
\examples{
sorting_speed(sample(1:10000), c("bubble", "insertion", "selection", "quick"))

}
